+++ packages/react-components/modals/src/BaseDialog/index.tsx
@@ -25,6 +25,10 @@ interface DialogProps {
    */
   additionalContent?: React.ReactNode;
   /**
+   * The background color of the dialog
+   */
+  backgroundColor?: string;
+  /**
    * The content of the BaseDialog
    */
   children: React.ReactNode;
@@ -49,7 +53,7 @@ interface DialogProps {
    */
   isOpen: boolean;
   /**
-   * function that is executed immediately after opening the modal
+   * Function that is executed immediately after opening the modal
    */
   onAfterOpen?: () => void;
   /**
@@ -57,6 +61,10 @@ interface DialogProps {
    * @param origin The original source of the request to close.
    */
   onClose: (origin: CloseOrigin) => void;
+  /**
+   * Class name applied directly to modal's portal
+   */
+  portalClassName?: string;
   /**
    * When set to true the Esc key can be used to close the dialog
    */
@@ -75,6 +83,7 @@ export type CloseOrigin = 'overlayClick' | 'escKey' | 'closeButton';
 
 const BaseDialog: React.FC<DialogProps> = ({
   additionalContent,
+  backgroundColor = tokens.colors.white,
   children,
   closeButtonDisabled = false,
   contentLabel,
@@ -83,6 +92,7 @@ const BaseDialog: React.FC<DialogProps> = ({
   isOpen,
   onAfterOpen,
   onClose,
+  portalClassName,
   shouldCloseOnEsc,
   shouldCloseOnOverlayClick,
   width,
@@ -123,6 +133,7 @@ const BaseDialog: React.FC<DialogProps> = ({
             onAfterOpen={onAfterOpen}
             onRequestClose={onRequestClose}
             overlayClassName={overlayClassName}
+            portalClassName={portalClassName}
             shouldCloseOnEsc={shouldCloseOnEsc}
             shouldCloseOnOverlayClick={shouldCloseOnOverlayClick}
           >
@@ -139,6 +150,7 @@ const BaseDialog: React.FC<DialogProps> = ({
             )}
             <Card
               css={{
+                backgroundColor,
                 display: 'flex',
                 flexDirection: 'column',
                 maxHeight: '100%',