+++ packages/docs/next-doc-site/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0-beta.15](https://github.com/datacamp/design-system/compare/@datacamp/waffles-next-doc-site@0.10.0-beta.14...@datacamp/waffles-next-doc-site@0.10.0-beta.15) (2022-07-29)
+
+**Note:** Version bump only for package @datacamp/waffles-next-doc-site
+
+
+
+
+
+# [0.10.0-beta.14](https://github.com/datacamp/design-system/compare/@datacamp/waffles-next-doc-site@0.10.0-beta.13...@datacamp/waffles-next-doc-site@0.10.0-beta.14) (2022-07-11)
+
+**Note:** Version bump only for package @datacamp/waffles-next-doc-site
+
+
+
+
+
+# [0.10.0-beta.13](https://github.com/datacamp/design-system/compare/@datacamp/waffles-next-doc-site@0.10.0-beta.12...@datacamp/waffles-next-doc-site@0.10.0-beta.13) (2022-05-17)
+
+**Note:** Version bump only for package @datacamp/waffles-next-doc-site
+
+
+
+
+
 # [0.10.0-beta.12](https://github.com/datacamp/design-system/compare/@datacamp/waffles-next-doc-site@0.10.0-beta.11...@datacamp/waffles-next-doc-site@0.10.0-beta.12) (2022-05-13)
 
 **Note:** Version bump only for package @datacamp/waffles-next-doc-site