+++ packages/react-components/modals/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-beta.11](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/compare/@datacamp/waffles-modals@8.0.0-beta.10...@datacamp/waffles-modals@8.0.0-beta.11) (2022-07-29)
+
+
+### Features
+
+* **dialog:** add backgroundColor prop ([c533cd8](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/commit/c533cd8))
+
+
+
+
+
+# [8.0.0-beta.10](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/compare/@datacamp/waffles-modals@8.0.0-beta.9...@datacamp/waffles-modals@8.0.0-beta.10) (2022-07-11)
+
+
+### Features
+
+* add portalClassName to BaseDialog and Dialog ([bea8456](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/commit/bea8456))
+
+
+
+
+
+# [8.0.0-beta.9](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/compare/@datacamp/waffles-modals@8.0.0-beta.8...@datacamp/waffles-modals@8.0.0-beta.9) (2022-05-17)
+
+**Note:** Version bump only for package @datacamp/waffles-modals
+
+
+
+
+
 # [8.0.0-beta.8](https://github.com/datacamp-engineering/design-system/tree/master/packages/react-components/modals/compare/@datacamp/waffles-modals@8.0.0-beta.7...@datacamp/waffles-modals@8.0.0-beta.8) (2022-05-13)
 
 **Note:** Version bump only for package @datacamp/waffles-modals